requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
httpx==0.27.0
gtfs-realtime-bindings==1.0.0
prefect==3.1.4
geopandas==1.0.1
shapely==2.0.6
asyncpg==0.30.0
pydantic==2.10.4
slowapi==0.1.9
celery==5.3.0
redis==4.5.1